refactor cookie list components
